#4 Updated integration-test-app to Grails 3.3.0
