+++ .gitignore
@@ -9,3 +9,4 @@
 .idea
 build
 cobertura.ser
+out
