+++ integration-test-app/build.gradle
@@ -1,25 +1,22 @@
 buildscript {
-	ext {
-		grailsVersion = project.grailsVersion
-	}
 	repositories {
 		mavenLocal()
-		maven { url 'https://repo.grails.org/grails/core' }
+		maven { url "https://repo.grails.org/grails/core" }
 	}
 	dependencies {
 		classpath "org.grails:grails-gradle-plugin:$grailsVersion"
+		classpath "org.grails.plugins:hibernate5:${gormVersion-".RELEASE"}"
 	}
 }
 
-plugins {
-	id 'io.spring.dependency-management' version '0.5.4.RELEASE'
-}
+version "0.1"
+group "integration.test.app"
 
-apply plugin: 'spring-boot'
 apply from: '../gradle/testapp.gradle'
 
 dependencies {
-	compile 'org.grails.plugins:hibernate'
+	compile "org.grails.plugins:hibernate5"
+	runtime "org.apache.tomcat:tomcat-jdbc"
 }
 
 integrationTest {
