+++ integration-test-app/gradle.properties
@@ -1,1 +1,4 @@
-grailsVersion=3.0.11
+grailsVersion=3.3.0
+grailsWrapperVersion=1.0.0
+gormVersion=6.1.6.RELEASE
+gradleWrapperVersion=3.5
