+++ integration-test-app/grails-app/conf/application.yml
@@ -1,32 +1,45 @@
 ---
 grails:
-   profile: web
-   spring:
-      transactionManagement: false
+    profile: web
+    codegen:
+        defaultPackage: integration.test.app
+    spring:
+        transactionManagement:
+            proxies: false
+    gorm:
+        reactor:
+            # Whether to translate GORM events into Reactor events
+            # Disabled by default for performance reasons
+            events: false
 info:
-   app:
-      name: '@info.app.name@'
-      version: '@info.app.version@'
-      grailsVersion: '@info.app.grailsVersion@'
+    app:
+        name: '@info.app.name@'
+        version: '@info.app.version@'
+        grailsVersion: '@info.app.grailsVersion@'
 spring:
-   groovy:
-      template:
-         check-template-location: false
+    main:
+        banner-mode: "off"
+    groovy:
+        template:
+            check-template-location: false
+
+# Spring Actuator Endpoints are Disabled by Default
+endpoints:
+    enabled: false
+    jmx:
+        enabled: true
 
 ---
 hibernate:
-   cache:
-      queries: false
-      use_query_cache: false
-      use_second_level_cache: false
-   format_sql: true
-   use_sql_comments: true
-
+    cache:
+        queries: false
+        use_second_level_cache: false
+        use_query_cache: false
 dataSource:
-   dbCreate: update
-   driverClassName: org.h2.Driver
-   jmxExport: false
-   password:
-   pooled: true
-   url: jdbc:h2:mem:testDb;MVCC=TRUE;LOCK_TIMEOUT=10000;DB_CLOSE_ON_EXIT=FALSE
-   username: sa
+    pooled: true
+    jmxExport: true
+    driverClassName: org.h2.Driver
+    username: sa
+    password: ''
+    dbCreate: create-drop
+    url: jdbc:h2:mem:devDb;MVCC=TRUE;LOCK_TIMEOUT=10000;DB_CLOSE_ON_EXIT=FALSE
