+++ integration-test-app/grails-app/conf/spring/resources.groovy
@@ -1,0 +1,6 @@
+import integration.test.app.UserPasswordEncoderListener
+
+// Place your Spring DSL code here
+beans = {
+    userPasswordEncoderListener(UserPasswordEncoderListener, ref('hibernateDatastore'))
+}
