+++ integration-test-app/grails-app/domain/integration/test/app/Permission.groovy
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test
+package integration.test.app
 
 /**
  * @author <a href='mailto:burt@burtbeckwith.com'>Burt Beckwith</a>
